requirements.txt
numpy==1.19.4
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
